% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BBCmsg}
\alias{BBCmsg}
\title{BBC articles}
\format{
A list of 4 character vectors containing the vectorized and stemmed documents
(i.e., unigrams with repetition) ;
 \describe{
  \item{msg1}{the birth of princess Charlotte}
  \item{msg2}{black holes in astrophysics}
  \item{msg3}{UK politics}
  \item{msg4}{cancer diseases in medicine}
}
}
\usage{
data("BBCmsg")
}
\description{
Pre-processed BBC articles from the BBC news network.
}
\keyword{datasets}
