% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{benchmark.gmm_lda}
\alias{benchmark.gmm_lda}
\title{Fit a Q-GMM in a K-LDA's latent space (Rmixmod)}
\usage{
benchmark.gmm_lda(dtm, Q, K, nruns = 1, seed = NULL)
}
\description{
Fit a Q-GMM in a K-LDA's latent space (Rmixmod)
}
