% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{benchmark.htsclust}
\alias{benchmark.htsclust}
\title{Clustering of count data via the Poisson mixture model of Rau et. al.}
\usage{
benchmark.htsclust(dtm.full, Q, ...)
}
\arguments{
\item{...}{: params to pass onto the PoisMixClus function of the HTSCluster package}
}
\description{
Clustering of count data via the Poisson mixture model of Rau et. al.
}
