% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{benchmark.kmeans_lda}
\alias{benchmark.kmeans_lda}
\title{Cluster the matrix theta obtained by a topicmodels LDA with K.est topics}
\usage{
benchmark.kmeans_lda(dtm.full, Q, K, ...)
}
\description{
Cluster the matrix theta obtained by a topicmodels LDA with K.est topics
}
