% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{benchmark.mpca_gmm}
\alias{benchmark.mpca_gmm}
\title{Fit a Q-GMM in a K-LDA's latent space (Rmixmod)}
\usage{
benchmark.mpca_gmm(dtm, Q, K, nruns = 1, seed = NULL)
}
\description{
Fit a Q-GMM in a K-LDA's latent space (Rmixmod)
}
