% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{benchmark.multmix}
\alias{benchmark.multmix}
\title{Clustering with the mixture of multinomials models}
\usage{
benchmark.multmix(dtm.full, Q, nruns)
}
\description{
Clustering with the mixture of multinomials models
}
