% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{benchmark.nmfem}
\alias{benchmark.nmfem}
\title{NMFEM algo from carel et. al.
MLE in a frequentist setting of MMPCA model}
\usage{
benchmark.nmfem(dtm, Q, K, nruns)
}
\description{
NMFEM algo from carel et. al.
MLE in a frequentist setting of MMPCA model
}
