% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{benchmarks-functions}
\alias{benchmarks-functions}
\alias{benchmark.htsclust}
\alias{benchmark.nmf}
\alias{benchmark.LDA}
\alias{benchmark.kmeans_lda}
\alias{benchmark.kmeans_dtm}
\alias{benchmark.multmix}
\alias{benchmark.nmfem}
\alias{benchmark.gmm_lda}
\title{Benchmarks functions for clustering}
\usage{
benchmark.htsclust(dtm.full, Q, ...)

benchmark.nmf(dtm.full, Q, ...)

benchmark.LDA(dtm.full, Q, ...)

benchmark.kmeans_lda(dtm.full, Q, K, ...)

benchmark.kmeans_dtm(dtm.full, Q, ...)

benchmark.multmix(dtm.full, Q, nruns)

benchmark.nmfem(dtm, Q, K, nruns)

benchmark.gmm_lda(dtm, Q, K, nruns = 1, seed = NULL)
}
\arguments{
\item{dtm.full}{}

\item{...}{: params to pass onto the PoisMixClus function of the HTSCluster package}

\item{nruns}{}

\item{seed}{}

\item{...}{}

\item{...}{}

\item{...}{}

\item{...}{}

\item{nruns}{}
}
\description{
These are wrapper to other methods for the clustering of count
  data. They can be use
}
\note{
zExper
}
\section{benchmark.htsclust}{

Clustering of count data via the Poisson mixture model of Rau et. al.
}

\section{\code{benchmarks.nmf}}{

Perform NMF document clustering of Lee and Seung .
}

\section{\code{benchmarks.kmeans_lda}}{

Cluster the matrix theta obtained by a topicmodels LDA with K.est topics
}

\section{\code{benchmarks.kmeans_dtm}}{

Naive Clustering on the dtm directly
}

\section{benchmarks.gmm_lda}{

dsncs
sdcknsqc

sdlkvnsd
}

