% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{DTMtoSparse}
\alias{DTMtoSparse}
\title{convert a dtm from package tm to sparseMatrix from package Matrix
without converting it to full matrix.}
\usage{
DTMtoSparse(dtm)
}
\arguments{
\item{dtm}{a document-term-matrix from package 'tm'}
}
\value{
a sparse dcgMatrix from package Matrix
}
\description{
convert a dtm from package tm to sparseMatrix from package Matrix
without converting it to full matrix.
}
