% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_functions.R
\name{initializeBeta}
\alias{initializeBeta}
\title{Function to calculate Beta on a given corpus
Several choices are available}
\usage{
initializeBeta(dtm, init.beta, K, verbose, control_lda_init)
}
\value{
a KxV matrix that which rows sums to 1
}
\description{
Function to calculate Beta on a given corpus
Several choices are available
}
