% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMPCA.R
\docType{package}
\name{MMPCA}
\alias{MMPCA}
\alias{MMPCA-package}
\title{MMPCA: Clustering for counts data}
\description{
The MMPCA package enables the clustering of counts data such as document/term matrix with
the mixture of multinomial PCA model. Model selection is performed via an
approximated form of the Integrated Classification Likelihood. The clustering
is done via the branch & bound described in the paper of ???.
}
