% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_and_bound_cvem.R
\name{mmpca_clust}
\alias{mmpca_clust}
\title{Perform clustering of count data using the MMPCA model.}
\usage{
mmpca_clust(dtm, Q, K, model = NULL, Yinit = "random",
  method = "CVEM", init.beta = "lda", keep = 1L, max.epochs = 100L,
  verbose = 1L)
}
\arguments{
\item{dtm}{an NxV \code{\link[tm]{DocumentTermMatrix}} with term-frequency
weighting.}

\item{Q}{The number of clusters}

\item{K}{The numer of topics (latent space dimension)}

\item{model}{A given model in which to take the controls for the VE-steps in
the greedy procedure. If NULL, a model of class \code{\linkS4class{mmpcaClust}} is
created with default controls (see \code{\linkS4class{mmpcaClustcontrol}}
class for more details).}

\item{Yinit}{Parameter for the initialization of Y. Tt can be either:
\itemize{ \item a string or a function specifying the initialization
procedure. It should be one of ('random', 'kmeans_lda'). See
\code{\link{benchmarks-functions}} functions for more details. \item A
vector of length N with Q modalities, specifying the initialization
clustering. }}

\item{method}{The clustering algorithm to be used. Currently, only "CVEM" is
available.}

\item{init.beta}{Parameter for the initialization of the matrix beta. It can
be either: \itemize{ \item a string specifying the initialization
procedure. It should be one of ('random', 'lda'). See
\code{\link{initializeBeta}}() for more details. \item A KxV matrix with
each row summing to 1.}}

\item{keep}{The evolution of the bound is tracked every \code{keep} iteration}

\item{max.epochs}{Specifies the maximum number of pass allowed on the whole
dataset.}

\item{verbose}{verbosity level}
}
\value{
An object of class \code{"\linkS4class{mmpcaClust}"} containing the
  fitted model.
}
\description{
Perform clustering of count data using the MMPCA model.
}
