% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmpcaClustcontrol_class.R
\docType{class}
\name{mmpcaClustcontrol-class}
\alias{mmpcaClustcontrol-class}
\title{mmpcaClustcontrol}
\description{
An S4 class for \code{\link{mmpca_clust}}(). It is mainly a
  wrapper around the class \code{\linkS4class{TopicModelcontrol}}
  (specifically: LDA_VEMcontrol).
}
\section{Slots}{

\describe{
\item{\code{control_lda_init}}{Object of class \code{"LDA_VEMcontrol"}; specifies the
controls of the VEM algorithm used for the initialization of beta.}

\item{\code{control_lda_loop}}{Object of class \code{"LDA_VEMcontrol"}; specifies the
controls for the VEM algorithm used after a swap in the branch & bound.}
}}

