% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoMPCA.R
\docType{package}
\name{MoMPCA}
\alias{MoMPCA}
\alias{MoMPCA-package}
\title{MoMPCA: Greedy clustering of count data through a mixture of multinomial PCA}
\description{
The MMPCA package implements the branch & bound classification-Variational
Expectation Maximisation algorithm described in Jouvin et. al.
\url{https://arxiv.org/abs/1909.00721}. It enables the clustering of counts
data such as document/term matrix modeled as a mixture of multinomial PCA
model. Model selection is performed via an approximated form of the
Integrated Classification Likelihood. .
}
\details{
The main entry point is the \code{\link{mmpca_clust}}() function to perfom
the clustering.
}
