% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot,mmpcaClust,missing-method}
\alias{plot,mmpcaClust,missing-method}
\title{Plot function for object mmpcaClust}
\usage{
\S4method{plot}{mmpcaClust,missing}(x, type = "topics", ...)
}
\arguments{
\item{x}{an S4 object of class \code{\linkS4class{mmpcaClust}}}

\item{type}{Either: \itemize{\item 'topics' (default): Show the top topic
words of topic matrix. See \code{\link{plot_topics}} documentation for more details. \item 'bound': plot the lower bound evolution during
the greedy procedure. See \code{\link{plot_bound}} documention for more details.}}

\item{...}{optional argument specifying the number of words to display and
the entropy correction to apply when calling \code{plot_topics}().}
}
\value{
a plot
}
\description{
Use ggplot2 if available.
}
