% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_bound}
\alias{plot_bound}
\title{Bound evolution plot}
\usage{
plot_bound(res)
}
\arguments{
\item{res}{An S4 object of class \code{\linkS4class{mmpcaClust}}}
}
\value{
a ggplot2 object if ggplot2 is available. Plot on the device
  otherwise.
}
\description{
Plot lower bound evolution
}
