% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_topics}
\alias{plot_topics}
\title{plot_topics}
\usage{
plot_topics(res, s = 2, n_words = 10)
}
\arguments{
\item{res}{An S4 object of class \code{\linkS4class{mmpcaClust}}}

\item{s}{an entropy correction parameter for the topic matrix. It is applied
to the beta matrix before sorting the words by highest probability. The
greater, the more emphasis is put towards words contributing a lot to the
entropy of a topic. Set s=1 to ignore.}

\item{n_words}{the number of words to display per topic.}
}
\value{
a ggplot2 object
}
\description{
Plot topic matrix
}
