% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{wrapper.mmpca_clust}
\alias{wrapper.mmpca_clust}
\title{Wrapper on mmpca_clust for several restart}
\usage{
wrapper.mmpca_clust(nruns = 5, Q.est, K.est)
}
\arguments{
\item{nruns}{number of restart of the algo}

\item{Q.est}{The number of cluster}

\item{K.est}{The number of topics}

\item{parrallel}{True if a parrallel or False}
}
\description{
Returns the best of all runs, i.e. the one wich achieves the greatest bound.
It works with the foreach package, if a parralel backend is registered it will run
the different initializations in parrallel.
}
\examples{
Wrapper function on mLDA to evaluate it on BBC (or other data set) with several restarts
}
